Extract components and parameters, add validation
File 'src/mininova.js' contains an app-agnostic description of all
parameters of the MiniNova.

Thus, component 'Controls' don't need to hold this information anymore.

A new 'Control' component will take care of rendering the appropriate
control, depending on its type (range or enum). Also, it will validate
that the value is contained in the range/enum.

The 'Oscillator' component delegates all rendering to 'Control', but
it's still possible to bypass this abstraction and render a more
specialized component.