Detect incoming NRPN messages
More complicated than expected, because the NRPN messages sent by the
MiniNova can be either 3 or 4 bytes long, depending on the control.

I have read that there could be an end sequence to clean previously set
MSB/LSB - like sending CC,0x63,0x7F and CC,0x62,0x7F - but this is not
the case with the MiniNova, so there is no easy way to distinguish
between 3 byte and 4 bytes long NRPN messages.

For example, changing the patch will trigger a 4 bytes long NRPN: 3rd
byte (value) will hold the bank number, and 4th byte (adjustment) will
hold the patch number.

On the contrary, animate pads will trigger a 3 bytes long NRPN, with
a single value hold by the 3rd byte.

As a result, I decided to hold a list of MSB/LSB tuples that identify
controls coded on 4 bytes. This list is specific to the MiniNova,
obviously.